+++ packages/sdk-client/src/api/api-client-options-helper.ts
@@ -29,7 +29,7 @@ export const buildOAuth2ApiClientOptions = (params: SinchClientParameters, apiNa
   return apiClientOptions;
 };
 
-/** @internal */
+/** @internal @deprecated */
 export const buildMailgunApiClientOptions = (
   params: Partial<MailgunCredentials & ApiPlugins & WithLogger>,
 ): ApiClientOptions => {